pyproject.toml
[project]
name = "fallacy-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas==2.2.3",
    "pyarrow==17.0.0",
    "torch>=2.10.0",
    "numpy>=1.24.0",
    "datasets",
    "ollama",
    "google-cloud-aiplatform>=1.142.0",
    "python-dotenv>=1.2.2",
    "z3-solver>=4.16.0.0",
    "google-genai>=1.68.0",
    "sentence-transformers>=5.4.1",
]
